fix: 修复 ps 和 memstat 工具的输出问题
问题修复:
- fill_memstat: 从 free lists 实际计算空闲页数
- sysproc.c: pstat 结构体 sz 字段类型改为 uint (与用户态一致)
- printf.c: 添加格式化宽度支持 (%-4d, %7s 等)
- ps.c: 简化输出逻辑，添加边界检查

+++ kernel/sysproc.c
@@ -224,7 +224,7 @@ struct pstat {
   int priority;
   int tickets;
   int sched_class;
-  uint64 sz;
+  uint sz;
   int rutime;
   int retime;
   int stime;
